requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
SQLAlchemy==2.0.34
alembic==1.13.2
python-dotenv==1.0.1
email-validator==2.2.0
itsdangerous==2.2.0
requests==2.32.3
gunicorn==23.0.0
